Add autoscaling feature to Vault's dynamodb table
To achieve this, we've had to create the dynamodb table from Terraform 
(before it was Vault itself that created the table on the first start). 
In theory that shouldn't be a problem, as the table structure is quite 
simple and future proof, and it can't be changed by Vault in any case. 
So we considered it safe to put in TF.

The autoscaling feature is optional (on by default).