Add autoscaling feature to Vault's dynamodb table (#3)
* Add autoscaling feature to Vault's dynamodb table

To achieve this, we've had to create the dynamodb table from Terraform 
(before it was Vault itself that created the table on the first start). 
In theory that shouldn't be a problem, as the table structure is quite 
simple and future proof, and it can't be changed by Vault in any case. 
So we considered it safe to put in TF.

The autoscaling feature is optional (on by default).

* Allow to override the dynamodb table name

This should make it easier to roll this out to the currently running 
setups.
Ref: 
https://github.com/skyscrapers/terraform-vault/pull/3/files#r203965604

* Fix dynamodb autoscaling target

* Allow to enable Letsencrypt staging server (for testing)

* Add a basic example and automated tests with terratest

* Fix le_staging variable in cloud-init template

* Use test environment in tests

* Setup nat gateway in test

* Add key_name as a variable in the test

* Complete automated test for a basic scenario

It sets up Vault, initializes it, and unseals every node, while checking 
their expected statuses

* Ignore terratest temp directory

* dep init and ignore vendor dir

* Add documentation on tests and configure r53 zone and acm via env vars

* Fix comments in test

* Add upgrade notice from v1.x to v2.x in README

* Update Variable documentation

* Added Ci task file to run the tests

* Install dep when running tests in ci

* Set ci input in gopath

* Missed to change dir in last commit

* Export gopath

* Install git in ci tests

* Change test domain name in examples to a generic one

* Added Letsencrypt email as a variable